feat: implement file deletion with system access check and add deleteById method in FileJPAGateway

+++ application/src/main/java/com/callv2/drive/application/file/delete/DefaultDeleteFileUseCase.java
@@ -1,6 +1,9 @@
 package com.callv2.drive.application.file.delete;
 
+import java.util.Objects;
+
 import com.callv2.drive.domain.exception.InternalErrorException;
+import com.callv2.drive.domain.exception.NotAllowedException;
 import com.callv2.drive.domain.exception.NotFoundException;
 import com.callv2.drive.domain.file.File;
 import com.callv2.drive.domain.file.FileGateway;
@@ -10,12 +13,11 @@ import com.callv2.drive.domain.member.MemberGateway;
 import com.callv2.drive.domain.member.MemberID;
 import com.callv2.drive.domain.storage.StorageService;
 
-import java.util.Objects;
-
 public class DefaultDeleteFileUseCase extends DeleteFileUseCase {
     private final MemberGateway memberGateway;
     private final FileGateway fileGateway;
     private final StorageService storageService;
+
     public DefaultDeleteFileUseCase(
             final MemberGateway memberGateway,
             final FileGateway fileGateway,
@@ -24,6 +26,7 @@ public class DefaultDeleteFileUseCase extends DeleteFileUseCase {
         this.fileGateway = Objects.requireNonNull(fileGateway);
         this.storageService = Objects.requireNonNull(storageService);
     }
+
     @Override
     public void execute(DeleteFileInput input) {
         final MemberID ownerId = MemberID.of(input.ownerId());
@@ -32,13 +35,13 @@ public class DefaultDeleteFileUseCase extends DeleteFileUseCase {
         final Member member = memberGateway.findById(ownerId)
                 .orElseThrow(() -> NotFoundException.with(Member.class, input.ownerId()));
 
-        member.hasSystemAccess();
+        if (!member.hasSystemAccess())
+            throw NotAllowedException.with("Member does not have permission to delete files");
 
         final File file = fileGateway.findById(fileId)
                 .orElseThrow(() -> NotFoundException.with(File.class, input.fileId().toString()));
 
-        fileGateway.deleteById(file.getId());
-
+        fileGateway.deleteById(file.getId()); // TODO maybe needs transactional
         deleteContentFile(file.getContent().location());
     }
